+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased changes
 
+## v1.10.2 (12 aug 2021)
+* [Issue 993](https://github.com/AmpersandTarski/Ampersand/issues/993) Don't call API to save field when value is unchanged
+* Provide more specific error message when file upload failed (e.g. exceeded maximum filesize)
+
 ## v1.10.1 (14 jul 2021)
 * Library update indicated by dependabot
 * Add helper methods to Population class
@@ -13,7 +17,7 @@
 * Update backend libraries (composer update)
 
 ## v1.9.3 (14 apr 2021)
-* [1155](https://github.com/AmpersandTarski/Ampersand/issues/1155) Fix issue with MrgAtom exec-engine function in case of non-existing atoms
+* [Issue 1155](https://github.com/AmpersandTarski/Ampersand/issues/1155) Fix issue with MrgAtom exec-engine function in case of non-existing atoms
 * Hotfix bug introduced in v1.9.2 related to query data optimization
 
 ## v1.9.2 (13 apr 2021)